libc: Remove compiler warnings.
This removes all compiler warnings related to:

  - Unused parameters.
  - Unused variables.
  - Type-punned pointer casts that break strict aliasing.

Note that the warnings about -std=gnu99 for C++ files persist.
The only reasonable way to handle these would be to do one of
these instead:

  - Hack the build system to remove the flag when it builds
    a C++ file.

  - Modify Android.mk to build all C++ files in a separate
    library, which doesn't use -std=gnu99

+++ libc/bionic/realpath.c
@@ -109,7 +109,7 @@ realpath(const char * __restrict path, char * __restrict resolved)
 		 */
 		p = strchr(left, '/');
 		s = p ? p : left + left_len;
-		if (s - left >= sizeof(next_token)) {
+		if (s - left >= (int)sizeof(next_token)) {
 			if (m)
 				free(resolved);
 			errno = ENAMETOOLONG;
@@ -201,7 +201,7 @@ realpath(const char * __restrict path, char * __restrict resolved)
 			 */
 			if (p != NULL) {
 				if (symlink[slen - 1] != '/') {
-					if (slen + 1 >= sizeof(symlink)) {
+					if (slen + 1 >= (int)sizeof(symlink)) {
 						if (m)
 							free(resolved);
 						errno = ENAMETOOLONG;
